# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Interactive Fiction Platform API
  description: OpenAPI 3.0 Specification for Interactive Fiction Platform (DAG Novel System)
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /api/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  description: "ขั้นต่ำ 8 ตัวอักษร ต้องมีตัวอักษรและตัวเลขอย่างละอย่างน้อย 1 ตัว"
                profileImage:
                  type: string
                  format: binary
                  description: "รูปโปรไฟล์ผู้ใช้ (ไม่บังคับ)"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refresh_token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/refresh:
    post:
      summary: Refresh authentication token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/users:
    get:
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/me/novels:
    get:
      summary: Get novels authored by current user (Writer)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of writer novels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Novel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /novels:
    get:
      summary: List all novels
      parameters:
        - in: query
          name: category_id
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: List of novels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Novel'
    post:
      summary: Create new novel (Writer)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, category_id]
              properties:
                title:
                  type: string
                description:
                  type: string
                category_id:
                  type: integer
                cover_image:
                  type: string
      responses:
        '201':
          description: Novel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Novel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /chapters:
    post:
      summary: Create new chapter (Writer)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id, title]
              properties:
                novel_id:
                  type: integer
                title:
                  type: string
                order_index:
                  type: integer
      responses:
        '201':
          description: Chapter created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /scenes:
    post:
      summary: Create new scene (Writer)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chapter_id, title, content]
              properties:
                chapter_id:
                  type: integer
                title:
                  type: string
                content:
                  type: string
                image_url:
                  type: string
                is_ending:
                  type: boolean
                ending_type:
                  type: string
                  enum: [Good Ending, Bad Ending, True Ending, Secret Ending]
                ending_title:
                  type: string
                ending_description:
                  type: string
      responses:
        '201':
          description: Scene created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scene'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /choices:
    post:
      summary: Create choice connecting from_scene to to_scene (DAG Edge)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from_scene_id, to_scene_id, choice_text]
              properties:
                from_scene_id:
                  type: integer
                to_scene_id:
                  type: integer
                choice_text:
                  type: string
      responses:
        '201':
          description: Choice created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Choice'
        '400':
          description: Bad request or DAG loop conflict detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /progress:
    post:
      summary: Record/Update reader current scene progress
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id, current_scene_id]
              properties:
                novel_id:
                  type: integer
                current_scene_id:
                  type: integer
      responses:
        '200':
          description: Reading progress recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /history:
    get:
      summary: Get reader history list (Read-Only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Reader history list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReadingHistory'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user-endings:
    post:
      summary: Record unlocked ending for reader
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id, scene_id]
              properties:
                novel_id:
                  type: integer
                scene_id:
                  type: integer
      responses:
        '200':
          description: Ending unlocked successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
  /bookshelves:
    get:
      summary: Get user bookshelf or check novel bookmark count
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: novel_id
          schema:
            type: integer
      responses:
        '200':
          description: Bookshelf details
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Add novel to bookshelf (Bookmark)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id]
              properties:
                novel_id:
                  type: integer
      responses:
        '201':
          description: Bookmarked successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Remove novel from bookshelf
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id]
              properties:
                novel_id:
                  type: integer
      responses:
        '200':
          description: Bookmark removed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
  /likes:
    post:
      summary: Like a novel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id]
              properties:
                novel_id:
                  type: integer
      responses:
        '200':
          description: Novel liked
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Unlike a novel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id]
              properties:
                novel_id:
                  type: integer
      responses:
        '200':
          description: Novel unliked
        '401':
          $ref: '#/components/responses/Unauthorized'
  /comments:
    post:
      summary: Add a comment to a novel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id, content]
              properties:
                novel_id:
                  type: integer
                scene_id:
                  type: integer
                content:
                  type: string
      responses:
        '201':
          description: Comment created
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Remove a comment
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: comment_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Comment deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource Conflict Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [reader, writer, admin]
...
    AuthResponse:
      type: object
      description: "Response จาก Login และ Register"
      properties:
        token:
          type: string
          description: "JWT Access Token"
        refresh_token:
          type: string
          description: "JWT Refresh Token"
        user:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
            email:
              type: string
            pic_profile:
              type: string
            role:
              type: string
              enum: [reader, writer, admin]
...